% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ryx.R
\name{plot.ryx}
\alias{plot.ryx}
\title{Plot function}
\usage{
\method{plot}{ryx}(x)
}
\arguments{
\item{x}{ryx}
}
\value{
a graph
}
\description{
Plot function
}
\examples{
library(MASS)
x <- ryx(Boston, y="medv")
plot(x)


}
