% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ryx.R
\name{print.ryx}
\alias{print.ryx}
\title{Print function}
\usage{
\method{print}{ryx}(x, ...)
}
\arguments{
\item{x}{ryx}

\item{...}{}
}
\value{
df
}
\description{
Print function
}
\examples{
library(MASS)
x <- ryx(Boston, y = "medv")
print(x)
}
