% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW.4 code finished.R
\name{ryx}
\alias{ryx}
\title{ryx function}
\usage{
ryx(data, y, x)
}
\arguments{
\item{data}{dataframe}

\item{y}{dependent variable}

\item{x}{independent variables}
}
\value{
variables and their correlations, p values, and stars against dependent variable
}
\description{
ryx function
}
\examples{

library(MASS)
x <- ryx(Boston, y="medv")
ryx(Boston, "medv")
}
