% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.ryx.R
\name{summary.ryx}
\alias{summary.ryx}
\title{Summary Function}
\usage{
\method{summary}{ryx}(x, ...)
}
\arguments{
\item{x}{ryx}
}
\value{
text
}
\description{
Summary Function
}
\examples{
library(MASS)
x <- ryx(Boston, y="medv")
summary(x)

}
